# docs/swagger.yaml
basePath: /
definitions:
  http.PostRequestRegisterAndAuth:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  http.PostRequestTask:
    properties:
      code:
        type: string
      tranlator:
        description: В тесте опечатка)))))))))
        type: string
    type: object
  http.PostResponseTask:
    properties:
      task_id:
        type: string
    type: object
  http.TaskResultResponse:
    properties:
      result:
        type: string
    type: object
  http.TaskStatusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a task service.
  title: Swagger Example API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Принимает JSON с данными username и password, сохраняет в куках
        айди сессии
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.PostRequestRegisterAndAuth'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Залогинить пользователя
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: Принимает JSON с данными username и password
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.PostRequestRegisterAndAuth'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарегестрировать пользователя
      tags:
      - User
  /result/{task_id}:
    get:
      description: Возвращает результат задачи по её UUID
      parameters:
      - description: UUID задачи
        format: uuid
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TaskResultResponse'
        "400":
          description: неверный формат UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить результат
      tags:
      - tasks
  /status/{task_id}:
    get:
      description: Возвращает текущий статус задачи по её UUID
      parameters:
      - description: UUID задачи
        format: uuid
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TaskStatusResponse'
        "400":
          description: неверный формат UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить статус
      tags:
      - tasks
  /task:
    post:
      consumes:
      - application/json
      description: Принимает JSON с данными для обработки и возвращает ID таски
      parameters:
      - description: Данные для таски
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.PostRequestTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.PostResponseTask'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать таску
      tags:
      - tasks
swagger: "2.0"